refacatoring trialcards
